+++ crates/prelude/CHANGELOG.md
@@ -31,4 +31,4 @@
 
 ## 0.1.0
 
-<!-- Update PR number to skip CHANGELOG.md test: #0 -->
+<!-- Update PR number to skip CHANGELOG.md test: #131 -->